Move QR code modal methods to their own page object
