+++ e2e/pages/Receive/RequestPaymentModal.js
